# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "List users",
        "tags": [
          "Users"
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Create user",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get user by ID",
        "tags": [
          "Users"
        ]
      },
      "put": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Update user",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Delete user",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/suppliers": {
      "get": {
        "operationId": "SuppliersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "List suppliers",
        "tags": [
          "Suppliers"
        ]
      },
      "post": {
        "operationId": "SuppliersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSupplierDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Create supplier",
        "tags": [
          "Suppliers"
        ]
      }
    },
    "/api/suppliers/{id}": {
      "get": {
        "operationId": "SuppliersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get supplier by ID",
        "tags": [
          "Suppliers"
        ]
      },
      "put": {
        "operationId": "SuppliersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSupplierDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Update supplier",
        "tags": [
          "Suppliers"
        ]
      },
      "delete": {
        "operationId": "SuppliersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Delete supplier",
        "tags": [
          "Suppliers"
        ]
      }
    }
  },
  "info": {
    "title": "BillBhai API",
    "description": "The BillBhai Retail Order Processing System API Documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "x-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role"
      }
    },
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "companyId": {
            "type": "string",
            "example": "BIZ-101",
            "description": "Company ID"
          },
          "name": {
            "type": "string",
            "example": "Arjun Mehta",
            "description": "User name"
          },
          "role": {
            "type": "string",
            "enum": [
              "superuser",
              "admin",
              "cashier",
              "returnhandler",
              "inventorymanager",
              "deliveryops",
              "customer"
            ],
            "description": "User role"
          },
          "email": {
            "type": "string",
            "example": "admin@billbhai.com",
            "description": "User email"
          },
          "mobileNo": {
            "type": "string",
            "example": "9870011201",
            "description": "Mobile number"
          },
          "username": {
            "type": "string",
            "example": "admin",
            "description": "Username"
          },
          "password": {
            "type": "string",
            "example": "admin123",
            "description": "Password"
          }
        },
        "required": [
          "companyId",
          "name",
          "role",
          "email",
          "mobileNo",
          "username",
          "password"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Arjun Mehta",
            "description": "User name"
          },
          "role": {
            "type": "string",
            "enum": [
              "superuser",
              "admin",
              "cashier",
              "returnhandler",
              "inventorymanager",
              "deliveryops",
              "customer"
            ],
            "description": "User role"
          },
          "email": {
            "type": "string",
            "example": "admin@billbhai.com",
            "description": "User email"
          },
          "mobileNo": {
            "type": "string",
            "example": "9870011201",
            "description": "Mobile number"
          },
          "password": {
            "type": "string",
            "example": "newpassword123",
            "description": "Password"
          },
          "status": {
            "type": "string",
            "enum": [
              "Active",
              "Inactive",
              "Suspended"
            ],
            "description": "User status"
          }
        }
      },
...
      "CreateSupplierDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Agarwal Traders",
            "description": "Supplier name"
          },
          "mobileNo": {
            "type": "string",
            "example": "9811544101",
            "description": "Mobile number"
          },
          "email": {
            "type": "string",
            "example": "agarwal@traders.in",
            "description": "Email address"
          },
          "address": {
            "type": "string",
            "example": "Delhi",
            "description": "Address"
          },
          "gstNo": {
            "type": "string",
            "example": "GSTSUP001",
            "description": "GST number"
          }
        },
        "required": [
          "name",
          "mobileNo",
          "email",
          "address"
        ]
      },
      "UpdateSupplierDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Agarwal Traders",
            "description": "Supplier name"
          },
          "mobileNo": {
            "type": "string",
            "example": "9811544101",
            "description": "Mobile number"
          },
          "email": {
            "type": "string",
            "example": "agarwal@traders.in",
            "description": "Email address"
          },
          "address": {
            "type": "string",
            "example": "Delhi",
            "description": "Address"
          },
          "gstNo": {
            "type": "string",
            "example": "GSTSUP001",
            "description": "GST number"
          }
        }
      }
    }
  },
  "security": [
    {
      "x-role": []
    }
  ]
}